refactor(settings): move docker-allowlist accessors and validator to settings_docker_allowlist.go (PRD #1021 M3)
